Connect to the WebSocket server from App/saga.js
Set the WebSocket protocol from the URL hash, which the server uses when
broadcasting messages to connected clients. Show a helpful message if
the URL hash is not set to help users get started.

The saga creates a WebSocket and a saga eventChannel to convert the
WebSocket events into redux actions. This allows us to use take*() in
other sagas to be able to respond to WebSocket messages. The
webSocketSendMessage() action is provided to allow sending an outgoing
message.