Add signaling to start connecting to remote client
Send a "start" message to the remote client when starting the
RemoteMedia saga. Likewise, when receiving "start" from the remote
client, we create an RTCPeerConnection, add the camera and microphone
tracks from UserMedia, and create an eventChannel that we can use to
feed events from the peer connection into Redux. The peer connection
should be closed when the remote client disconnects via its channel,
since RTCPeerConnection does not have a "close" event of its own.