Resolve "start" signaling conflicts using a timestamp and a nonce
Both clients send a "start" message, so it is possible for both clients
to create an offer. This is problematic because WebRTC expects one side
to be the offering side, while the other is the answering side. Resolve
"start" conflicts using a timestamp and a nonce in the "start" payload.
The client that sent "start" first wins the collision and creates the
offer.

To facilitate local development with hot loading, use a random nonce as
well to be able to resolve the conflict if both clients send "start"
with the exact same timestamp.